Switched to ListView object for Icons.  It is displaying them properly.
TODO:

Update folders in the view to have the functionality of files (will need to update the XML to accomplish this.

Icon Form:
Add Buttons:
Add Icon:
When clicked fileselectdialog opens and allows multiple files to be selected. (filter only shows .ico files) When they are selected confirmation popup box says "Adding Icons:
iconname
" Are you sure (yes/no)

Delete Icon:
When cliccked popup box says "Delete Icon iconname"
Are you sure (Yes/no)"

When an icon is clicked a textbox with the name of it (which is the filename without the ext) is populated with the name.  Type something in the box or have it be updated different than the name of the icon, then a update button appears next to it.
Click update says "you are updating icon from BLAH to blah," Are you sure (yes/no)