Icon changing in the Icons Form works reliably if only changing 1 icon and then saving and closing after.  It does do more than 1 icon without saving / closing but it can glitch  with  out of memory issues and file in use issues.
TODO: see if removal of memory cache of icons step fixes the having to save / out of memory after each change.

Need to start working on Adding icon and deleting.

If permissions / file access issues are still happening as of EOD 9/30/21 then I should switch to a "working" icons directory that is created each time the program is launched and then written to each time the program is exited.  (at least for now so I can finish the rest of the 3 open issues relating to icons.