Issues to fix and where to pick up:
Icons management and displaying them for folders:

mRemoteNG being the default Icon for folders:
When an object is created regardless of whether it's a folder or a connection it ends up with icon field being populated with mRemoteNG.  So to fix folders having mRemoteNG as the default icon, I will have to find where the icon value is being set to mRemoteNG when an object is created and either do an if statement based on if it's a container (if that value is available), or something similar at some point after the point where it's set to mRemoteNG.

Still to do:

1.  Remove any memory caching / usage for storing the icons and confirm it doesn't affect the applications functions.
2.  Complete the rest of the Icon Settings functions.
a.  Have change icon button either say save update icon (or something like that) or create new button to update the name of the icon (which would rename the icon file to name.ico).
b.  Delete Icon.
c.  Add Icon.

3.  Once all functionality works and there are no real issues:

a.  Compare the changes with the latest development branch and remove anything that is not needed.
b.  confirm there are descriptive notes on each change.
c.  get the latest development branch and modify it with the changes so only one merge request (which uses the latest commit as it's source) is generated.
d.  Write up the changes using the repo form for commits, use previous commits as the format for the details.
e.  do merge request.